move resourceTypeAndCredentialManager to packages/proxy
