+++ packages/proxy/lib/http/util/cookies.ts
@@ -2,8 +2,8 @@ import _ from 'lodash'
 import type Debug from 'debug'
 import { URL } from 'url'
 import { SerializableAutomationCookie, Cookie, CookieJar, toughCookieToAutomationCookie } from '@packages/server/lib/util/cookies'
-import type { RequestCredentialLevel } from '../../types'
-import type { ResourceType } from '@packages/net-stubbing'
+import type { RequestCredentialLevel, ResourceType } from '../../resourceTypeAndCredentialManager'
+
 import { urlOriginsMatch, urlSameSiteMatch } from '@packages/network-tools'
 
 type SiteContext = 'same-origin' | 'same-site' | 'cross-site'
