+++ packages/proxy/lib/index.ts
@@ -1,3 +1,5 @@
 export { NetworkProxy } from './network-proxy'
 
 export * from './types'
+
+export * from './resourceTypeAndCredentialManager'
