+++ packages/server/lib/server-base.ts
@@ -33,7 +33,6 @@ import type { FoundSpec, ProtocolManagerShape, TestingType } from '@packages/typ
 import type { Server as WebSocketServer } from 'ws'
 import { RemoteStates, RemoteState } from './remote_states'
 import { cookieJar, SerializableAutomationCookie } from './util/cookies'
-import { resourceTypeAndCredentialManager, ResourceTypeAndCredentialManager } from './util/resourceTypeAndCredentialManager'
 import * as fileServer from './file_server'
 import type { FileServer } from './file_server'
 import appData from './util/app_data'
@@ -46,6 +45,7 @@ import type Protocol from 'devtools-protocol'
 import type { ServiceWorkerClientEvent } from '@packages/proxy/lib/http/util/service-worker-manager'
 import type { Automation } from './automation'
 import type { AutomationCookie } from './automation/cookies'
+import type { ResourceType, RequestCredentialLevel } from '@packages/proxy'
 
 const debug = Debug('cypress:server:server-base')
 
@@ -151,7 +151,6 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
   protected request: Request
   protected isListening: boolean
   protected socketAllowed: SocketAllowed
-  protected resourceTypeAndCredentialManager: ResourceTypeAndCredentialManager
   protected _fileServer: FileServer | null
   protected _baseUrl: string | null
   protected _server?: DestroyableHttpServer
@@ -190,8 +189,6 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
     }
 
     this._remoteStates = new RemoteStates(remoteStatePorts, this._documentDomainInjection)
-
-    this.resourceTypeAndCredentialManager = resourceTypeAndCredentialManager
   }
 
   ensureProp = ensureProp
@@ -242,7 +239,9 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
       this.socket.toDriver('cross:origin:cookies', cookies)
     })
 
-    this.socket.localBus.on('request:sent:with:credentials', this.resourceTypeAndCredentialManager.set)
+    this.socket.localBus.on('request:sent:with:credentials', (credentials: { url: string, resourceType: ResourceType, credentialStatus: RequestCredentialLevel }) => {
+      this._networkProxy?.setCredentials(credentials)
+    })
   }
 
   createServer (
@@ -359,7 +358,6 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
     this.createNetworkProxy({
       config,
       remoteStates: this._remoteStates,
-      resourceTypeAndCredentialManager: this.resourceTypeAndCredentialManager,
       shouldCorrelatePreRequests,
       getCurrentBrowser,
     })
@@ -449,7 +447,7 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
     return e
   }
 
-  createNetworkProxy ({ config, remoteStates, resourceTypeAndCredentialManager, shouldCorrelatePreRequests, getCurrentBrowser }) {
+  createNetworkProxy ({ config, remoteStates, shouldCorrelatePreRequests, getCurrentBrowser }) {
     const getFileServerToken = () => {
       return this._fileServer?.token
     }
@@ -466,7 +464,6 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
       netStubbingState: this.netStubbingState,
       request: this.request,
       serverBus: this._eventBus,
-      resourceTypeAndCredentialManager,
       getCurrentBrowser,
     })
   }
@@ -484,7 +481,7 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
       this.networkProxy.reset({ resetBetweenSpecs: false })
       this.netStubbingState.reset()
       this._remoteStates.reset()
-      this.resourceTypeAndCredentialManager.clear()
+      this.networkProxy.clearCredentials()
     }
 
     const ios = this.socket.startListening(this.server, automation, config, options)
@@ -647,7 +644,6 @@ export class ServerBase<TSocket extends SocketE2E | SocketCt> {
 
   reset () {
     this._networkProxy?.reset({ resetBetweenSpecs: true })
-    this.resourceTypeAndCredentialManager.clear()
     const baseUrl = this._baseUrl ?? '<root>'
 
     return this._remoteStates.set(baseUrl)
